refactor: extract admin KPI cards into reusable component
